+++ src/types/effect.rs
@@ -1,38 +1,78 @@
 //! Contains effects via the [`Effect`] type.
 
-use crate::types::iterators::FlatteningIntoIterator;
 use crate::types::CEffect;
+use std::ops::Deref;
 
 /// An effect. Occurs e.g. in a [`ActionDefinition`](crate::types::ActionDefinition).
 ///
+/// This represents the `(and ...)` variant of the PDDL definition,
+/// modeling cases of zero, one or many effects.
+///
 /// ## Usage
 /// Used by [`ActionDefinition`](crate::ActionDefinition) and [`CEffect`].
-#[derive(Debug, Clone, PartialEq)]
-pub enum Effect<'a> {
-    /// Exactly the specified effect applies.
-    Single(CEffect<'a>), // TODO: Unify with `All` variant; this is just a single-element vector and according to spec, this vector may be empty.
-    /// Conjunction: All effects apply (i.e. a and b and c ..).
-    All(Vec<CEffect<'a>>),
-}
+#[derive(Debug, Clone, PartialEq, Default)]
+pub struct Effect<'a>(Vec<CEffect<'a>>);
 
-impl<'a> IntoIterator for Effect<'a> {
-    type Item = CEffect<'a>;
-    type IntoIter = FlatteningIntoIterator<Self::Item>;
+impl<'a> Effect<'a> {
+    /// Constructs a new instance from the value.
+    pub fn new(effect: CEffect<'a>) -> Self {
+        Self(vec![effect])
+    }
 
-    fn into_iter(self) -> Self::IntoIter {
-        match self {
-            Effect::Single(item) => FlatteningIntoIterator::new(item),
-            Effect::All(vec) => FlatteningIntoIterator::new_vec(vec),
+    /// Constructs a new instance from the provided vector of values.
+    pub fn new_and(effects: Vec<CEffect<'a>>) -> Self {
+        Self(effects)
+    }
+
+    /// Returns `true` if the list contains no elements.
+    pub fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
+
+    /// Returns the number of elements in the list, also referred to
+    /// as its 'length'.
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    /// Returns an iterator over the list.
+    ///
+    /// The iterator yields all items from start to end.
+    pub fn iter(&'a self) -> std::slice::Iter<'a, CEffect<'a>> {
+        self.0.iter()
+    }
+
+    /// Get the only element of this list if the list has
+    /// exactly one element. Returns [`None`] in all other cases.
+    pub fn try_get_single(self) -> Option<CEffect<'a>> {
+        if self.len() == 1 {
+            self.into_iter().next()
+        } else {
+            None
         }
     }
 }
 
-impl<'a> Effect<'a> {
-    pub const fn new(effect: CEffect<'a>) -> Self {
-        Self::Single(effect)
+impl<'a> IntoIterator for Effect<'a> {
+    type Item = CEffect<'a>;
+    type IntoIter = std::vec::IntoIter<Self::Item>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.into_iter()
     }
-    pub fn new_and<I: IntoIterator<Item = CEffect<'a>>>(effect: I) -> Self {
-        Self::All(effect.into_iter().collect())
+}
+
+impl<'a> Deref for Effect<'a> {
+    type Target = [CEffect<'a>];
+
+    fn deref(&self) -> &Self::Target {
+        self.0.as_slice()
+    }
+}
+
+impl<'a> AsRef<[CEffect<'a>]> for Effect<'a> {
+    fn as_ref(&self) -> &[CEffect<'a>] {
+        self.0.as_slice()
     }
 }
 
@@ -50,6 +90,14 @@ impl<'a> From<Vec<CEffect<'a>>> for Effect<'a> {
 
 impl<'a> FromIterator<CEffect<'a>> for Effect<'a> {
     fn from_iter<T: IntoIterator<Item = CEffect<'a>>>(iter: T) -> Self {
-        Effect::new_and(iter)
+        Effect::new_and(iter.into_iter().collect())
+    }
+}
+
+impl<'a> TryInto<CEffect<'a>> for Effect<'a> {
+    type Error = ();
+
+    fn try_into(self) -> Result<CEffect<'a>, Self::Error> {
+        self.try_get_single().ok_or(())
     }
 }